requirements/base.txt
django==4.2.1
psycopg2-binary==2.9.6
python-environ==0.4.54
django-cors-headers==4.0.0
django-jazzmin==2.6.0
drf-yasg==1.21.5
djangorestframework==3.14.0
django-filter==23.2
django-rosetta==0.9.9
python-telegram-bot==13.15  # for telegram bot integration
django-image-uploader-widget==0.2.5
django-ckeditor==6.5.1
celery==5.2.7
redis==4.5.5
django-modeltranslation==0.18.9
django-recaptcha==3.0.0
django-auditlog==2.2.2
django-rosetta==0.9.9
django-phonenumbers==1.0.1
django-phonenumber-field==7.1.0
Pillow==9.5.0
Unidecode==1.3.6
qrcode==7.4.2
weasyprint==59.0
django-resized==1.0.2